+++ src/SoundFingerprinting/Utils/FastFingerprintDescriptor.cs
@@ -6,7 +6,7 @@
 
         public override bool[] ExtractTopWavelets(float[] frames, int topWavelets)
         {
-            int[] indexes = RangeUtils.GetRange(frames.Length);
+            ushort[] indexes = RangeUtils.GetRange(frames.Length);
             quickSelect.Find(topWavelets - 1, frames, indexes, 0, frames.Length - 1);
             return EncodeFingerprint(frames, indexes, topWavelets);
         }
