+++ src/SoundFingerprinting/Utils/QuickSelectAlgorithm.cs
@@ -2,7 +2,7 @@
 {
     public class QuickSelectAlgorithm
     {
-        public int Find(int kth, float[] list, int[] indexes, int lo, int hi)
+        public int Find(int kth, float[] list, ushort[] indexes, int lo, int hi)
         {
             while (lo != hi)
             {
@@ -30,7 +30,7 @@
             return lo;
         }
 
-        private int Partition(float[] list, int[] indexes, int pivotIndex, int lo, int hi)
+        private int Partition(float[] list, ushort[] indexes, int pivotIndex, int lo, int hi)
         {
             float pivotValue = Abs(list[pivotIndex]);
             Swap(list, indexes, pivotIndex, hi);
@@ -48,18 +48,18 @@
             return storeIndex;
         }
 
-        private void Swap(float[] list, int[] indexes, int i, int j)
+        private void Swap(float[] list, ushort[] indexes, int i, int j)
         {
             float tmp = list[i];
             list[i] = list[j];
             list[j] = tmp;
 
-            int indexTmp = indexes[i];
+            ushort indexTmp = indexes[i];
             indexes[i] = indexes[j];
             indexes[j] = indexTmp;
         }
 
-        private void SwapIfGreater(float[] list, int[] indexes, int a, int b)
+        private void SwapIfGreater(float[] list, ushort[] indexes, int a, int b)
         {
             if (a != b)
             {
